Refactor
Rewritten to FSD architechture
