--- src/components/place-card/index.tsx
@@ -1,2 +1,0 @@
-export { default as MainPlaceCard } from './main-place-card';
-export { default as FavoritesPlaceCard } from './favorites-place-card';
