+++ src/entities/offer/index.ts
@@ -1,0 +1,1 @@
+export type { OfferType } from './types';
