+++ src/entities/review/index.ts
@@ -1,0 +1,2 @@
+export { type ReviewType } from './types';
+export { default as Review } from './ui/review';
