+++ src/entities/review/types.ts
@@ -1,0 +1,11 @@
+export type ReviewType = {
+  id: string;
+  date: string;
+  user: {
+    name: string;
+    avatarUrl: string;
+    isPro: boolean;
+  };
+  comment: string;
+  rating: number;
+};
