+++ src/features/favorites/favorite/index.tsx
@@ -1,0 +1,2 @@
+export * from './ui';
+
