+++ src/features/favorites/favorite/model/index.tsx
@@ -1,0 +1,1 @@
+export const addToFavorites = (id: string) => id;
