+++ src/features/favorites/favorite/ui/bookmark.tsx
@@ -1,0 +1,20 @@
+import { addToFavorites } from '../model';
+
+type PropsType = {
+  id: string;
+};
+
+const FavoriteButton = ({ id }: PropsType) => (
+  <button
+    onClick={() => addToFavorites(id)}
+    className="place-card__bookmark-button button"
+    type="button"
+  >
+    <svg className="place-card__bookmark-icon" width="18" height="19">
+      <use xlinkHref="#icon-bookmark"></use>
+    </svg>
+    <span className="visually-hidden">To bookmarks</span>
+  </button>
+);
+
+export default FavoriteButton;
