+++ src/features/favorites/favorite/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as FavoriteButton } from './bookmark';
