+++ src/features/review/send/index.tsx
@@ -1,0 +1,1 @@
+export * from './ui';
