+++ src/features/review/send/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as SendReviewForm } from './Form';
