+++ src/pages/offer/loader.ts
@@ -1,12 +1,14 @@
 import { LoaderFunctionArgs } from 'react-router-dom';
 
-import { Offer } from '../../types/offer';
+import { OfferType } from '../../entities/offer';
 
-import offers from '../../mocks/offer';
-import comments from '../../mocks/comment';
+import offers from '../../entities/offer/api/mock';
+import comments from '../../entities/review/api/mock';
 
 export const offerDataLoader = ({ params }: LoaderFunctionArgs) => {
-  const targetOffer = offers.find((offer) => offer.id === params.id) as Offer;
+  const targetOffer = offers.find(
+    (offer) => offer.id === params.id,
+  ) as OfferType;
   return {
     offer: targetOffer,
     comments,
