+++ src/shared/framework/index.ts
@@ -1,0 +1,1 @@
+export { default as PrivateRoute } from './private-route';
