+++ src/shared/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as BaseLayout } from './layout/base';
+export { default as Header } from './header';
