+++ src/shared/ui/layout/base.tsx
@@ -1,0 +1,12 @@
+import { Outlet } from 'react-router-dom';
+
+import { Header } from '..';
+
+const BaseLayout = () => (
+  <>
+    <Header />
+    <Outlet />
+  </>
+);
+
+export default BaseLayout;
