--- src/types/comments.ts
@@ -1,11 +1,0 @@
-export type Comment = {
-  id: string;
-  date: string;
-  user: {
-    name: string;
-    avatarUrl: string;
-    isPro: boolean;
-  };
-  comment: string;
-  rating: number;
-};
