+++ src/widgets/place/ui/index.tsx
@@ -1,0 +1,3 @@
+export { default as MainPlaceCard } from './main-place-card';
+export { default as FavoritesPlaceCard } from './favorites-place-card';
+export { default as MainPlacesList } from './main-places-list';
